web_app/pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = "A simple Flask web application for a task management system."
authors = ["Juan Esteban Garcia <juanesgarciamar@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
Flask-SQLAlchemy = "^3.1.1"
flask-marshmallow = "^1.2.1"
psycopg2 = "^2.9.9"
marshmallow-sqlalchemy = "^1.0.0"
python-dotenv = "^1.0.1"
Flask-WTF = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
